startLevel and targetLevel refactored from Integer to Double

+++ test/controllers/admin/metrics/SaveTest.java
@@ -180,8 +180,8 @@ public class SaveTest extends ControllerTest<Save> {
     controller.infoSource = "http://";
     controller.institutionToReport = "f   ";
     controller.goalId = 5L;
-    controller.startLevel = 5;
-    controller.targetLevel = 6;
+    controller.startLevel = 5.0;
+    controller.targetLevel = 6.0;
     Criteria criteria = mock(Criteria.class);
     when(hibernate.createCriteria(Metric.class)).thenReturn(criteria);
     when(criteria.createCriteria(anyString())).thenReturn(criteria);
@@ -201,8 +201,8 @@ public class SaveTest extends ControllerTest<Save> {
     assertEquals("http://", savedMetric.getInfoSource());
     assertEquals("f", savedMetric.getInstitutionToReport());
     assertEquals((Double) 8.0, savedMetric.getOrderNumber());
-    assertEquals(5, (int) savedMetric.getStartLevel());
-    assertEquals(6, (int) savedMetric.getTargetLevel());
+    assertEquals((Double) 5.0,savedMetric.getStartLevel());
+    assertEquals((Double) 6.0,savedMetric.getTargetLevel());
     assertEquals(false, savedMetric.getIsPublic());
   }
 
@@ -242,16 +242,16 @@ public class SaveTest extends ControllerTest<Save> {
     controller.unit = "%";
     controller.publicDescription = "a a a";
     controller.privateDescription = "b";
-    controller.startLevel = 5;
+    controller.startLevel = 5.0;
     controller.commentOnStartLevel = "c";
-    controller.targetLevel = 6;
+    controller.targetLevel = 6.0;
     controller.commentOnTargetLevel = "d";
     controller.infoSource = "http://";
     controller.institutionToReport = "f";
     controller.orderNumber = 5.0;
     controller.isPublic = true;
 
-    Metric metricBeingChanged = new Metric(new Goal("", 10), "TERE", null, null, null, 0, null, 0, null, null, null, 1.0, false);
+    Metric metricBeingChanged = new Metric(new Goal("", 10), "TERE", null, null, null, 0.0, null, 0.0, null, null, null, 1.0, false);
     when(hibernate.get(Metric.class, 2L)).thenReturn(metricBeingChanged);
 
     assertRender(controller.post());
@@ -261,9 +261,9 @@ public class SaveTest extends ControllerTest<Save> {
     assertEquals("%", updatedMetric.getUnit());
     assertEquals("a a a", updatedMetric.getPublicDescription());
     assertEquals("b", updatedMetric.getPrivateDescription());
-    assertEquals(5, (int) updatedMetric.getStartLevel());
+    assertEquals((Double) 5.0,updatedMetric.getStartLevel());
     assertEquals("c", updatedMetric.getCommentOnStartLevel());
-    assertEquals(6, (int) updatedMetric.getTargetLevel());
+    assertEquals((Double) 6.0, updatedMetric.getTargetLevel());
     assertEquals("d", updatedMetric.getCommentOnTargetLevel());
     assertEquals("http://", updatedMetric.getInfoSource());
     assertEquals("f", updatedMetric.getInstitutionToReport());
